Proxy support for Desktop version
Add proxy support for desktop version by replacing `fetch` with
`cross-fetch` in the Octokit constructor.

VS code already patches the node modules net, http, and https to
utilize its proxy settings. So to get proxies to work in this web
extension for the desktop version, octokit simply needs to use a fetch
implementation that uses the node components in a node environment.